+++ aeppl/__init__.py
@@ -14,6 +14,7 @@ from aeppl.printing import latex_pprint, pprint
 import aeppl.cumsum
 import aeppl.mixture
 import aeppl.scan
+import aeppl.tensor
 import aeppl.transforms
 import aeppl.truncation
